refactor(release): extract both manifest-version writers to release_env
Paying Phase 2's own source-size charge rather than bumping a ratchet.
Adding set_augment_manifest_version to release.ts cost +35 excess lines,
because release.ts sits 564 lines past the 1500-line threshold and every
line above it is counted.

Both writers — the new set_augment_manifest_version and its pre-existing
twin set_marketplace_version — move whole into release_env.ts, the leaf that
already owns both manifest paths AND jsonDumpsIndent, so neither needed
anything it did not already have there. release_env.ts is 298 lines, far
under the threshold. Splitting a matched pair across two files for no reason
would have been the worse shape, which is why the twin moved too.

release.ts re-exports set_marketplace_version so its existing test import
keeps resolving; the new writer is imported from release_env directly.

Net effect: 18,571 -> 18,566, five lines BELOW the previous baseline, so the
baseline is lowered to the exact live total (check_source_size_budget.test.ts
asserts equality) with the reason recorded in its note.

Also regenerates both metric rows against the post-merge tree (294 skills,
959 links, dead_links still empty, 24 dangles across 17 survivors unchanged)
and updates the census claim to name the merge rather than pin a stale count.
