+++ agents/evidence/metrics/skill-link-census.json
@@ -1,17 +1,17 @@
 {
   "schema_version": 1,
   "_comment": "Written by `./scripts-run src/scripts/lint_handoffs --census-json`. Do not hand-edit — regenerate it. The collector is `collect_body_links`, the same one the gate scans with, so the published figure and the gate verdict cannot drift apart. `dead_links` uses the gate predicate (a target absent from the live SKILL.md set), NOT the Reproduction B.1 grep — see `links_matched_by_gate` for why the two differ and which dead link the grep was blind to.",
-  "commit": "c7e82087e1402968b9ecf16d43cc3affd3968e47",
-  "total_links_skill_md": 941,
-  "total_links_any_dir": 954,
-  "links_matched_by_gate": 930,
-  "files_with_skill_md_link": 205,
-  "files_with_any_relative_link": 224,
-  "skills_total": 292,
+  "commit": "9c4f5bff4170c196ab3a1cfc612bde120252c5eb",
+  "total_links_skill_md": 959,
+  "total_links_any_dir": 972,
+  "links_matched_by_gate": 948,
+  "files_with_skill_md_link": 207,
+  "files_with_any_relative_link": 226,
+  "skills_total": 294,
   "dead_links": [],
-  "undeclared_in_requires": 922,
+  "undeclared_in_requires": 940,
   "skills_declaring_requires": 5,
-  "scoped_survivors": 221,
+  "scoped_survivors": 223,
   "scoped_pruned": 71,
   "scoped_dangles": [
     {
